backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery (Background Jobs)
celery==5.3.6
flower==2.0.1

# Vector Database
qdrant-client==1.7.3

# MinIO (S3)
minio==7.2.3
boto3==1.34.34

# AI & ML
google-generativeai==0.3.2
langchain==0.1.6
langchain-google-genai==0.0.6

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.4
python-docx==1.1.0
python-pptx==0.6.23
pillow==10.2.0
pytesseract==0.3.10
python-magic-bin==0.4.14
imagehash==4.3.1
pypdf==4.0.1

# Text Processing
charset-normalizer==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# LINE SDK
line-bot-sdk==3.7.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3
